Users: Improve interfaces for deleting users.
Provide clearer guidance for reassigning or deleting content when deleting users in both single and multisite instances. Content assignment is handled on a per-user basis, rather than forcing all content to be assigned to a single user. Add validation to enforce making a selection. Add fieldsets and legends to improve form structure and accessibility. Move inline JS to `common.js`. Improve backend logic to process per-user content handling. Exclude users being deleted from dropdown forms.

Developed in https://github.com/WordPress/wordpress-develop/pull/10502, https://github.com/WordPress/wordpress-develop/pull/12467

Props stefan.velthuys, ocean90, hellofromTonya, akshat2802, hubersen, fakhriaz, ozgursar, sirlouen, realloc, audrasjb, joedolson, afercia, krokodok.
Fixes #56914.
Built from https://develop.svn.wordpress.org/trunk@62688


git-svn-id: http://core.svn.wordpress.org/trunk@61972 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-admin/users.php
@@ -213,12 +213,17 @@ switch ( $wp_list_table->current_action() ) {
 				continue;
 			}
 
-			switch ( $_REQUEST['delete_option'] ) {
+			if ( 'reassign' === $_REQUEST['delete_option'][ $id ] && empty( $_REQUEST['reassign_user'][ $id ] ) ) {
+				$update = 'err_missing_reassign';
+				continue;
+			}
+
+			switch ( $_REQUEST['delete_option'][ $id ] ) {
 				case 'delete':
 					wp_delete_user( $id );
 					break;
 				case 'reassign':
-					wp_delete_user( $id, $_REQUEST['reassign_user'] );
+					wp_delete_user( $id, $_REQUEST['reassign_user'][ $id ] );
 					break;
 			}
 
@@ -306,40 +311,9 @@ switch ( $wp_list_table->current_action() ) {
 			$user_ids = array_diff( $user_ids, array( $current_user->ID ) );
 		}
 
-		/**
-		 * Filters whether the users being deleted have additional content
-		 * associated with them outside of the `post_author` and `link_owner` relationships.
-		 *
-		 * @since 5.2.0
-		 *
-		 * @param bool  $users_have_additional_content Whether the users have additional content. Default false.
-		 * @param int[] $user_ids                      Array of IDs for users being deleted.
-		 */
-		$users_have_content = (bool) apply_filters( 'users_have_additional_content', false, $user_ids );
-
-		if ( $user_ids && ! $users_have_content ) {
-			if ( $wpdb->get_var(
-				"SELECT ID FROM {$wpdb->posts}
-				WHERE post_author IN( " . implode( ',', $user_ids ) . ' )
-				LIMIT 1'
-			) ) {
-				$users_have_content = true;
-			} elseif ( $wpdb->get_var(
-				"SELECT link_id FROM {$wpdb->links}
-				WHERE link_owner IN( " . implode( ',', $user_ids ) . ' )
-				LIMIT 1'
-			) ) {
-				$users_have_content = true;
-			}
-		}
-
-		if ( $users_have_content ) {
-			add_action( 'admin_head', 'delete_users_add_js' );
-		}
-
 		require_once ABSPATH . 'wp-admin/admin-header.php';
 		?>
-		<form method="post" name="updateusers" id="updateusers">
+		<form method="post" name="updateusers" id="updateusers" class="delete-and-reassign-users-form">
 		<?php wp_nonce_field( 'delete-users' ); ?>
 		<?php echo $referer; ?>
 
@@ -365,7 +339,8 @@ switch ( $wp_list_table->current_action() ) {
 
 		<ul>
 		<?php
-		$go_delete = 0;
+		$go_delete          = 0;
+		$users_have_content = false;
 
 		foreach ( $all_user_ids as $id ) {
 			$user = get_userdata( $id );
@@ -385,12 +360,97 @@ switch ( $wp_list_table->current_action() ) {
 					'<input type="hidden" name="users[]" value="%s" />',
 					esc_attr( $id )
 				);
-				printf(
-					/* translators: 1: User ID, 2: User login. */
-					__( 'ID #%1$s: %2$s' ),
-					$id,
-					$user->user_login
-				);
+
+				/**
+				 * Filters whether the users being deleted have additional content
+				 * associated with them outside of the `post_author` and `link_owner` relationships.
+				 *
+				 * @since 5.2.0
+				 *
+				 * @param bool  $users_have_additional_content Whether the users have additional content. Default false.
+				 * @param int[] $user_ids                      Array of IDs for users being deleted.
+				 */
+				$user_has_content = (bool) apply_filters( 'users_have_additional_content', false, array( $id ) );
+
+				if ( ! $user_has_content ) {
+					if ( $wpdb->get_var(
+						$wpdb->prepare(
+							"SELECT ID FROM {$wpdb->posts}
+							WHERE post_author = %d
+							LIMIT 1",
+							$id
+						)
+					) ) {
+						$user_has_content = true;
+					} elseif ( $wpdb->get_var(
+						$wpdb->prepare(
+							"SELECT link_id FROM {$wpdb->links}
+							WHERE link_owner = %d
+							LIMIT 1",
+							$id
+						)
+					) ) {
+						$user_has_content = true;
+					}
+				}
+
+				if ( ! $user_has_content ) {
+					?>
+					<input type="hidden" name="delete_option[<?php echo esc_attr( $id ); ?>]" value="delete" required />
+					<p>
+					<?php
+					printf(
+						/* translators: 1: User login. 2: User ID */
+						__( '%1$s (ID #%2$s): This user does not have any content.' ),
+						'<strong>' . $user->user_login . '</strong>',
+						$id
+					);
+					?>
+					</p>
+					<?php
+				} else {
+					?>
+					<fieldset>
+					<legend>
+					<?php
+					printf(
+						/* translators: 1: User login, 2: User ID. */
+						__( '%1$s (ID #%2$s): What should be done with the content owned by this user?' ),
+						'<strong>' . $user->user_login . '</strong>',
+						$id
+					);
+					?>
+					</legend>
+
+					<ul>
+						<li>
+							<input type="radio" id="delete_option_<?php echo esc_attr( $id ); ?>" name="delete_option[<?php echo esc_attr( $id ); ?>]" value="delete" required />
+							<label for="delete_option_<?php echo esc_attr( $id ); ?>"><?php _e( 'Delete all content.' ); ?></label>
+						</li>
+						<li>
+							<input type="radio" id="reassign_option_<?php echo esc_attr( $id ); ?>" name="delete_option[<?php echo esc_attr( $id ); ?>]" value="reassign" required />
+							<label for="reassign_option_<?php echo esc_attr( $id ); ?>"><?php _e( 'Attribute all content to another user.' ); ?></label>
+
+							<label for="reassign_user_<?php echo esc_attr( $id ); ?>" class="screen-reader-text"><?php _e( 'Select a user to attribute the content to.' ); ?></label>
+							<?php
+							wp_dropdown_users(
+								array(
+									'show_option_none'  => __( 'Select a user' ),
+									'option_none_value' => '',
+									'name'              => 'reassign_user[' . $id . ']',
+									'id'                => 'reassign_user_' . $id,
+									'exclude'           => $user_ids,
+									'show'              => 'display_name_with_login',
+								)
+							);
+							?>
+						</li>
+					</ul>
+					</fieldset>
+					<?php
+					$users_have_content = true;
+				}
+
 				echo "</li>\n";
 
 				++$go_delete;
@@ -401,41 +461,6 @@ switch ( $wp_list_table->current_action() ) {
 
 		<?php
 		if ( $go_delete ) :
-
-			if ( ! $users_have_content ) :
-				?>
-				<input type="hidden" name="delete_option" value="delete" />
-			<?php else : ?>
-				<fieldset>
-				<?php if ( 1 === $go_delete ) : ?>
-					<p><legend><?php _e( 'What should be done with content owned by this user?' ); ?></legend></p>
-				<?php else : ?>
-					<p><legend><?php _e( 'What should be done with content owned by these users?' ); ?></legend></p>
-				<?php endif; ?>
-
-				<ul style="list-style:none;">
-					<li>
-						<input type="radio" id="delete_option0" name="delete_option" value="delete" />
-						<label for="delete_option0"><?php _e( 'Delete all content.' ); ?></label>
-					</li>
-					<li>
-						<input type="radio" id="delete_option1" name="delete_option" value="reassign" />
-						<label for="delete_option1"><?php _e( 'Attribute all content to:' ); ?></label>
-						<?php
-						wp_dropdown_users(
-							array(
-								'name'    => 'reassign_user',
-								'exclude' => $user_ids,
-								'show'    => 'display_name_with_login',
-							)
-						);
-						?>
-					</li>
-				</ul>
-				</fieldset>
-				<?php
-			endif;
-
 			/**
 			 * Fires at the end of the delete users form prior to the confirm button.
 			 *
@@ -612,10 +637,11 @@ switch ( $wp_list_table->current_action() ) {
 
 		$messages = array();
 		if ( isset( $_GET['update'] ) ) :
+			$delete_count = isset( $_GET['delete_count'] ) ? (int) $_GET['delete_count'] : 0;
+
 			switch ( $_GET['update'] ) {
 				case 'del':
 				case 'del_many':
-					$delete_count = isset( $_GET['delete_count'] ) ? (int) $_GET['delete_count'] : 0;
 					if ( 1 === $delete_count ) {
 						$message = __( 'User deleted.' );
 					} else {
@@ -695,14 +721,16 @@ switch ( $wp_list_table->current_action() ) {
 							'dismissible'        => true,
 						)
 					);
-					$messages[] = wp_get_admin_notice(
-						__( 'Other user roles have been changed.' ),
-						array(
-							'id'                 => 'message',
-							'additional_classes' => array( 'updated' ),
-							'dismissible'        => true,
-						)
-					);
+					if ( $delete_count > 0 ) {
+						$messages[] = wp_get_admin_notice(
+							__( 'Other user roles have been changed.' ),
+							array(
+								'id'                 => 'message',
+								'additional_classes' => array( 'updated' ),
+								'dismissible'        => true,
+							)
+						);
+					}
 					break;
 				case 'err_admin_del':
 					$messages[] = wp_get_admin_notice(
@@ -713,14 +741,36 @@ switch ( $wp_list_table->current_action() ) {
 							'dismissible'        => true,
 						)
 					);
+					if ( $delete_count > 0 ) {
+						$messages[] = wp_get_admin_notice(
+							__( 'Other users have been deleted.' ),
+							array(
+								'id'                 => 'message',
+								'additional_classes' => array( 'updated' ),
+								'dismissible'        => true,
+							)
+						);
+					}
+					break;
+				case 'err_missing_reassign':
 					$messages[] = wp_get_admin_notice(
-						__( 'Other users have been deleted.' ),
+						__( 'Users could not be deleted because no user was selected for content reassignment.' ),
 						array(
 							'id'                 => 'message',
-							'additional_classes' => array( 'updated' ),
+							'additional_classes' => array( 'error' ),
 							'dismissible'        => true,
 						)
 					);
+					if ( $delete_count > 0 ) {
+						$messages[] = wp_get_admin_notice(
+							__( 'Other users have been deleted.' ),
+							array(
+								'id'                 => 'message',
+								'additional_classes' => array( 'updated' ),
+								'dismissible'        => true,
+							)
+						);
+					}
 					break;
 				case 'remove':
 					$messages[] = wp_get_admin_notice(
